Eliminate bugs and make calculator work as expected
The calculator is working as expected. I added keyboard
support. However the implementation seems a bit
cumbersome and brute force since I targeted each button
of the calculator to correspond to matching keys
on the keyboard. I wished I found cleaner a way where
I populate the screen directly with the keyboard
while respecting all exceptions declared in the
event delegated switch cases in the populateDisplay()
and getOperator() functions.
For now, I'll have a little bit of styling to do.